+++ src/Plainion.GraphViz.Viewer/CoreModule.cs
@@ -1,12 +1,11 @@
 ﻿using System.ComponentModel.Composition;
 using Plainion.GraphViz.Viewer.Views;
-using Prism.Mef.Modularity;
+using Prism.Ioc;
 using Prism.Modularity;
 using Prism.Regions;
 
 namespace Plainion.GraphViz.Viewer
 {
-    [ModuleExport(typeof(CoreModule))]
     class CoreModule : IModule
     {
         private IRegionManager myRegionManager;
@@ -17,11 +16,15 @@ namespace Plainion.GraphViz.Viewer
             myRegionManager = regionManager;
         }
 
-        public void Initialize()
+        public void RegisterTypes(IContainerRegistry containerRegistry)
         {
             myRegionManager.RegisterViewWithRegion(GraphViz.Viewer.RegionNames.GraphViewer, typeof(GraphViewer));
             myRegionManager.RegisterViewWithRegion(GraphViz.Viewer.RegionNames.SettingsEditor, typeof(SettingsEditor));
             myRegionManager.RegisterViewWithRegion(GraphViz.Viewer.RegionNames.StatusMessagesViewer, typeof(StatusMessagesView));
+        }
+
+        public void OnInitialized(IContainerProvider containerProvider)
+        {
         }
     }
 }