+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 - Package analysis: fixed generating clusters for namespaces
 - Ensure process is killed when exception is throw during startup
+- Dependencies updated
 
 ## 3.3 - 2019-07-18