+++ ChangeLog.md
@@ -6,6 +6,7 @@
 - Cleanup of APIs of modules
 - "dot.exe" home no longer read from config (always use embedded distribution)
 - Switched to Unity as MEF is no longer supported in Prism 7
+- Migrated to Prism 7.2
 
 ## 3.3 - 2019-07-18